Move the fetch /push code from RepoActionsExt to gitbutler-git crate
