+++ crates/but-api/Cargo.toml
@@ -60,7 +60,6 @@ legacy = [
     "dep:gitbutler-reference",
     "dep:gitbutler-stack",
     "dep:gitbutler-repo",
-    "dep:gitbutler-repo-actions",
     "dep:gitbutler-edit-mode",
     "dep:gitbutler-operating-modes",
     "dep:gitbutler-oplog",
@@ -115,7 +114,6 @@ gitbutler-commit = { workspace = true, optional = true }
 gitbutler-reference = { workspace = true, optional = true }
 gitbutler-stack = { workspace = true, optional = true }
 gitbutler-repo = { workspace = true, optional = true }
-gitbutler-repo-actions = { workspace = true, optional = true }
 gitbutler-edit-mode = { workspace = true, optional = true }
 gitbutler-operating-modes = { workspace = true, optional = true }
 gitbutler-oplog = { workspace = true, optional = true }
