+++ crates/gitbutler-branch-actions/src/base.rs
@@ -9,10 +9,10 @@ use but_ctx::Context;
 use but_error::Marker;
 use but_oxidize::ObjectIdExt;
 use gitbutler_branch::GITBUTLER_WORKSPACE_REFERENCE;
+use gitbutler_git::GitContextExt as _;
 use gitbutler_project::FetchResult;
 use gitbutler_reference::{Refname, RemoteRefname};
 use gitbutler_repo::first_parent_commit_ids_until_with_graph;
-use gitbutler_repo_actions::RepoActionsExt;
 use gitbutler_stack::{Stack, Target, VirtualBranchesHandle, canned_branch_name};
 use serde::Serialize;
 use tracing::instrument;
