+++ crates/gitbutler-git/Cargo.toml
@@ -28,20 +28,27 @@ benches = []
 askpass = []
 
 [dependencies]
+but-askpass.workspace = true
+but-core.workspace = true
+but-ctx.workspace = true
+but-error.workspace = true
 tracing.workspace = true
 thiserror.workspace = true
 anyhow.workspace = true
+gitbutler-reference.workspace = true
 tokio = { workspace = true, optional = true, features = [
     "process",
     "time",
     "io-util",
     "net",
     "fs",
+    "rt-multi-thread",
 ] }
 rand.workspace = true
 futures.workspace = true
 sysinfo = { workspace = true, features = ["system"] }
 gix.workspace = true
+serde.workspace = true
 
 [target."cfg(unix)".dependencies]
 nix = { version = "0.30.0", features = ["process", "socket", "user"] }
