+++ crates/gitbutler-repo-actions/Cargo.toml
@@ -10,24 +10,13 @@ rust-version.workspace = true
 doctest = false
 
 [dependencies]
-but-error.workspace = true
-but-askpass.workspace = true
 but-ctx.workspace = true
 
 gitbutler-stack.workspace = true
-gitbutler-git.workspace = true
-gitbutler-reference.workspace = true
 gitbutler-repo.workspace = true
 
 gix.workspace = true
-tokio = { workspace = true, features = [
-    "rt-multi-thread",
-    "rt",
-    "macros",
-    "sync",
-] }
 anyhow.workspace = true
-tracing.workspace = true
 
 [lints]
 workspace = true
