+++ crates/gitbutler-stack/tests/stack.rs
@@ -15,8 +15,8 @@ use but_db::DbHandle;
 use but_meta::virtual_branches_legacy_types as legacy_types;
 use but_testsupport::{gix_testtools, open_repo};
 use filetime::{FileTime, set_file_mtime};
+use gitbutler_git::GitContextExt as _;
 use gitbutler_reference::RemoteRefname;
-use gitbutler_repo_actions::RepoActionsExt;
 use gitbutler_stack::{PatchReferenceUpdate, Stack, StackBranch, Target, VirtualBranchesHandle};
 use gix::refs::transaction::PreviousValue;
 use itertools::Itertools;
